requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
Automat==22.10.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
exceptiongroup==1.0.4
filelock==3.8.1
Flask==2.2.2
frozenlist==1.3.3
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
multidict==6.0.3
numpy==1.23.5
outcome==1.2.0
packaging==21.3
pandas==1.5.2
parsel==1.7.0
Protego==0.2.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
scrapy-selenium==0.0.7
selenium==4.7.2
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tldextract==3.4.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.0
webdriver-manager==3.8.5
Werkzeug==2.2.2
wsproto==1.2.0
yarl==1.8.2
zope.interface==5.5.2
